Update main to avoid warning

+++ src/main.m
@@ -25,12 +25,12 @@ int
 main (int argc, const char *argv[])
 {
   NSAutoreleasePool *pool = [[NSAutoreleasePool alloc] init];
-  NSUInteger i = 0;
+  int i = 0;
   
   // Process arguments, currently we ignore them but this avoids the warning.
   for (i = 0; i < argc; i++)
     {
-      printf("Argument#%ld - %s", i, argv[i]);
+      printf("Argument#%d - %s", i, argv[i]);
     }
   
   [ScoreMakerDocumentController sharedDocumentController];
